componentize SearchInputHeader
