+++ packages/apolloschurchapp/src/tabs/discover/Discover.js
@@ -1,57 +1,15 @@
 import React, { PureComponent } from 'react';
-import { Platform, StyleSheet, View } from 'react-native';
 import { Query } from 'react-apollo';
 import { get } from 'lodash';
 
-import {
-  FeedView,
-  BackgroundView,
-  styled,
-  PaddedView,
-  SearchInput,
-} from '@apollosproject/ui-kit';
+import { FeedView, BackgroundView } from '@apollosproject/ui-kit';
+
+import SearchInputHeader, {
+  ReactNavigationStyleReset,
+} from '../../ui/SearchInputHeader';
 
 import TileContentFeed from './TileContentFeed';
 import GET_CONTENT_CHANNELS from './getContentChannels';
-
-const HeaderBorder = styled(
-  ({ theme }) => ({
-    paddingBottom: 8,
-    /* It's unclear why this is necessary but without it the layout breaks on both platforms. Limited
-     * research suggest that without a background color the shadows don't know what to blend with so
-     * the view collapses. */
-    backgroundColor: theme.colors.background.paper,
-    // Renders the same shadows that React Navigation does.
-    ...Platform.select({
-      ios: {
-        shadowColor: 'rgba(0, 0, 0, 0.3)',
-        shadowOpacity: 0.85,
-        shadowRadius: 0,
-        shadowOffset: {
-          width: 0,
-          height: StyleSheet.hairlineWidth,
-        },
-      },
-      android: {
-        elevation: 4,
-      },
-    }),
-  }),
-  'SearchInputHeader.HeaderBorder'
-)(PaddedView);
-
-// This element is used to clip the Android shadow in every directection except the bottom.
-const ClipAndroidElevationFix = styled(
-  {
-    ...Platform.select({
-      android: {
-        paddingBottom: 4,
-        overflow: 'hidden',
-      },
-    }),
-  },
-  'SearchInputHeader.ClipAndroidElevationFix'
-)(View);
 
 const childContentItemLoadingState = {
   title: '',
@@ -62,13 +20,6 @@ const feedItemLoadingState = {
   name: '',
   isLoading: true,
 };
-
-const styles = StyleSheet.create({
-  header: {
-    borderBottomWidth: 0,
-    elevation: 0,
-  },
-});
 
 class Discover extends PureComponent {
   renderItem = ({ item }) => (
@@ -86,11 +37,7 @@ class Discover extends PureComponent {
   render() {
     return (
       <BackgroundView>
-        <ClipAndroidElevationFix>
-          <HeaderBorder vertical={false}>
-            <SearchInput />
-          </HeaderBorder>
-        </ClipAndroidElevationFix>
+        <SearchInputHeader />
         <Query query={GET_CONTENT_CHANNELS} fetchPolicy="cache-and-network">
           {({
             error,
@@ -116,7 +63,7 @@ class Discover extends PureComponent {
 
 Discover.navigationOptions = {
   title: 'Discover',
-  headerStyle: styles.header,
+  headerStyle: ReactNavigationStyleReset.header,
 };
 
 export default Discover;
